Handle error, don't do manual sort, info helper
Gracefully handle httperror, exit

On the YouTube Studio page changed the playlist not to be manually sort. This fixed the issue where a channel is added and all of its videos go to the top. Doesn't seem to be a way to modify the sort of the playlist using the API.

Created a function to get info about the playlist using the API. Used for debugging but left in the code